cloudrun_functions/webPage/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-cloud-storage
Flask
matplotlib